+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+* Send ctrl-C to child process first instead of always forcibly killing it.
+* Report command failure as a service-specific error to Windows.
+* Added `--stop-timeout` option.
+
 ## v0.1.0 (2019-09-22)
 
 * Initial release.